WorldFacts: Further workaround for iOS 5 VoiceOver bug
The table view method dequeueReusableCellWithIdentifier fails to
return a cell when VoiceOver is active. Falling back to a manual
load using UINib creates a new cell but has two impacts:

1) The new cell is not associated with the segue defined in the
storyboard. This means we need to detect these manually created
cells in tableView:didSelectRowAtIndexPath: and manually trigger
the segue. To ensure we only do this for these cells we use a
different reuseIdentifier when creating the cell with UINib.

2) Since the dequeue mechanism fails when VoiceOver is active the
new cells are never reused. This means that as the table is scrolled
or reloaded we continue to create new cells. The cells are small
but eventually this could lead to memory issues.