Update TwitterSearch to use Twitter API v1.1
The Twitter API v1 that was used by this project has now been
retired and no longer returns search results.

This project has now been updated to use the Social framework
so that it can use the Twitter credentials of the device user.
The minimum deployment target is now iOS 6.0.

I also added a refresh control to the table view controller to
refresh the search results.