Adapting stack view now works with iOS 9 beta 4
Commented out workaround to bug in early iOS 9 betas to
change stack view axis when view size changes. This is no
longer necessary as the stack view now correctly adapts
to size class changes.

Left the code commented as it provides an example of how
to react to size changes.