Updated Refresh for Xcode 8
This project dated back to iOS 6 so I gave it a bit of a spring
clean to remove some old cruft and refactor the view controller
code to have a separate data source.

The workaround for the storyboard bug that meant the UIRefreshControl
action was never called is also no longer needed.